chore: model types
