+++ src/leapfrogai_api/typedef/models/__init__.py
@@ -1,1 +1,5 @@
-from .model_types import Model as Model
+from .model_types import (
+    Model as Model,
+    ModelResponse as ModelResponse,
+    ModelResponseModel as ModelResponseModel,
+)
